+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### 🐛 Bug Fixes
 
+- Fix CSV export "to" range to include the whole day in all cases ([#12286](https://github.com/blockscout/blockscout/pull/12286))
 - Return compatibility with previous version of health endpoint([#12280](https://github.com/blockscout/blockscout/pull/12280))
 - Unbind import from compile-time chain_type ([#12277](https://github.com/blockscout/blockscout/pull/12277))
 - Read `CHAIN_TYPE` and `MUD_INDEXER_ENABLED` envs in runtime config ([#12270](https://github.com/blockscout/blockscout/issues/12270))